Nitpick SDK updates

+++ src/main/java/io/mailtrap/api/suppressions/SuppressionsImpl.java
@@ -7,7 +7,7 @@ import io.mailtrap.http.RequestData;
 import io.mailtrap.model.response.suppressions.SuppressionsResponse;
 
 import java.net.URLEncoder;
-import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,30 +15,34 @@ import static io.mailtrap.http.RequestData.entry;
 
 public class SuppressionsImpl extends ApiResource implements Suppressions {
 
-  public SuppressionsImpl(MailtrapConfig config) {
-    super(config);
-    this.apiHost = Constants.GENERAL_HOST;
-  }
+    public SuppressionsImpl(final MailtrapConfig config) {
+        super(config);
+        this.apiHost = Constants.GENERAL_HOST;
+    }
 
-  @Override
-  public List<SuppressionsResponse> search(long accountId, String email) {
-    var queryParams = RequestData.buildQueryParams(entry("email", Optional.ofNullable(email)));
+    @Override
+    public List<SuppressionsResponse> search(final long accountId, final String email) {
+        final var queryParams = RequestData.buildQueryParams(entry("email", Optional.ofNullable(email)));
 
-    return
-        httpClient.getList(
-            String.format(apiHost + "/api/accounts/%d/suppressions", accountId),
-            new RequestData(queryParams),
-            SuppressionsResponse.class
-        );
-  }
+        return
+            httpClient.getList(
+                String.format(apiHost + "/api/accounts/%d/suppressions", accountId),
+                new RequestData(queryParams),
+                SuppressionsResponse.class
+            );
+    }
 
-  @Override
-  public SuppressionsResponse deleteSuppression(long accountId, String suppressionId) {
-    return
-        httpClient.delete(
-            String.format(apiHost + "/api/accounts/%d/suppressions/%s", accountId, URLEncoder.encode(suppressionId, Charset.defaultCharset())),
-            new RequestData(),
-            SuppressionsResponse.class
-        );
-  }
+    @Override
+    public SuppressionsResponse deleteSuppression(final long accountId, final String suppressionId) {
+        if (suppressionId == null || suppressionId.isBlank()) {
+            throw new IllegalArgumentException("suppressionId must not be null or blank");
+        }
+
+        return
+            httpClient.delete(
+                String.format(apiHost + "/api/accounts/%d/suppressions/%s", accountId, URLEncoder.encode(suppressionId, StandardCharsets.UTF_8)),
+                new RequestData(),
+                SuppressionsResponse.class
+            );
+    }
 }
